+++ src/AppRouter.jsx
@@ -11,7 +11,7 @@ const AppRouter = () => {
         //ruta de login y registro
     <Route path="/auth/*" element={<AuthRouter/>}/>
         //ruta app
-    <Route path="/app" element={<JournalRouter/>}/>
+    <Route path="/app/*" element={<JournalRouter/>}/>
 </Routes>
